rename: core 모듈 내 repository 클래스들을 각 도메인 서비스 모듈로 이동
