+++ admin/src/main/java/kernel360/ckt/admin/application/DrivingLogService.java
@@ -8,8 +8,8 @@ import kernel360.ckt.admin.ui.dto.response.DrivingLogListResponse;
 import kernel360.ckt.core.domain.entity.DrivingLogEntity;
 import kernel360.ckt.core.domain.entity.RouteEntity;
 import kernel360.ckt.core.domain.enums.DrivingType;
-import kernel360.ckt.core.repository.DrivingLogRepository;
-import kernel360.ckt.core.repository.RouteRepository;
+import kernel360.ckt.admin.infra.DrivingLogRepository;
+import kernel360.ckt.admin.infra.RouteRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
