+++ admin/src/main/java/kernel360/ckt/admin/application/RentalService.java
@@ -12,11 +12,11 @@ import kernel360.ckt.core.domain.entity.DrivingLogEntity;
 import kernel360.ckt.core.domain.entity.RentalEntity;
 import kernel360.ckt.core.domain.entity.VehicleEntity;
 import kernel360.ckt.core.domain.enums.RentalStatus;
-import kernel360.ckt.core.repository.CompanyRepository;
-import kernel360.ckt.core.repository.CustomerRepository;
-import kernel360.ckt.core.repository.DrivingLogRepository;
-import kernel360.ckt.core.repository.RentalRepository;
-import kernel360.ckt.core.repository.VehicleRepository;
+import kernel360.ckt.admin.infra.CompanyRepository;
+import kernel360.ckt.admin.infra.CustomerRepository;
+import kernel360.ckt.admin.infra.DrivingLogRepository;
+import kernel360.ckt.admin.infra.RentalRepository;
+import kernel360.ckt.admin.infra.VehicleRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
