+++ admin/src/main/java/kernel360/ckt/admin/application/VehicleService.java
@@ -5,8 +5,7 @@ import jakarta.persistence.EntityNotFoundException;
 import jakarta.transaction.Transactional;
 import kernel360.ckt.admin.application.command.CreateVehicleCommand;
 import kernel360.ckt.admin.application.command.UpdateVehicleCommand;
-import kernel360.ckt.admin.infra.repository.jpa.RentalJpaRepository;
-import kernel360.ckt.admin.ui.dto.request.VehicleUpdateRequest;
+import kernel360.ckt.admin.infra.jpa.RentalJpaRepository;
 import kernel360.ckt.admin.ui.dto.response.ControlTowerSummaryResponse;
 import kernel360.ckt.admin.ui.dto.response.GpsPointResponse;
 import kernel360.ckt.admin.ui.dto.response.RunningVehicleResponse;
@@ -16,8 +15,8 @@ import kernel360.ckt.core.domain.entity.RentalEntity;
 import kernel360.ckt.core.domain.entity.VehicleEntity;
 import kernel360.ckt.core.domain.enums.RentalStatus;
 import kernel360.ckt.core.domain.enums.VehicleStatus;
-import kernel360.ckt.core.repository.RentalRepository;
-import kernel360.ckt.core.repository.VehicleRepository;
+import kernel360.ckt.admin.infra.RentalRepository;
+import kernel360.ckt.admin.infra.VehicleRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
