+++ admin/src/main/java/kernel360/ckt/admin/domain/projection/WeeklyVehicleLogProjection.java
@@ -1,4 +1,4 @@
-package kernel360.ckt.admin.infra.repository.projection;
+package kernel360.ckt.admin.domain.projection;
 
 import java.time.LocalDate;
