+++ admin/src/main/java/kernel360/ckt/admin/infra/RouteRepository.java
@@ -1,11 +1,9 @@
-package kernel360.ckt.core.repository;
+package kernel360.ckt.admin.infra;
 
 import kernel360.ckt.core.domain.entity.DrivingLogEntity;
 import kernel360.ckt.core.domain.entity.RouteEntity;
-import kernel360.ckt.core.domain.enums.RouteStatus;
 
 import java.util.List;
-import java.util.Optional;
 
 public interface RouteRepository {
     RouteEntity save(RouteEntity route);
