+++ admin/src/main/java/kernel360/ckt/admin/infra/jpa/RentalJpaRepository.java
@@ -1,4 +1,4 @@
-package kernel360.ckt.admin.infra.repository.jpa;
+package kernel360.ckt.admin.infra.jpa;
 
 import java.time.LocalDateTime;
 import java.util.List;
@@ -6,7 +6,7 @@ import java.util.Optional;
 import kernel360.ckt.core.domain.entity.RentalEntity;
 import kernel360.ckt.core.domain.entity.VehicleEntity;
 import kernel360.ckt.core.domain.enums.RentalStatus;
-import kernel360.ckt.core.repository.RentalRepository;
+import kernel360.ckt.admin.infra.RentalRepository;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.EntityGraph;
