+++ admin/src/main/java/kernel360/ckt/admin/infra/jpa/RouteJpaRepository.java
@@ -1,7 +1,7 @@
-package kernel360.ckt.admin.infra.repository.jpa;
+package kernel360.ckt.admin.infra.jpa;
 
 import kernel360.ckt.core.domain.entity.RouteEntity;
-import kernel360.ckt.core.repository.RouteRepository;
+import kernel360.ckt.admin.infra.RouteRepository;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface RouteJpaRepository extends JpaRepository<RouteEntity, Long>, RouteRepository {
