+++ admin/src/main/java/kernel360/ckt/admin/infra/jpa/VehicleJpaRepository.java
@@ -1,8 +1,8 @@
-package kernel360.ckt.admin.infra.repository.jpa;
+package kernel360.ckt.admin.infra.jpa;
 
 import kernel360.ckt.core.domain.entity.VehicleEntity;
 import kernel360.ckt.core.domain.enums.VehicleStatus;
-import kernel360.ckt.core.repository.VehicleRepository;
+import kernel360.ckt.admin.infra.VehicleRepository;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
